+++ tests/realtikvtest/addindextest3/functional_test.go
@@ -22,7 +22,6 @@ import (
 	"github.com/pingcap/tidb/pkg/ddl"
 	"github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/testkit"
-	"github.com/pingcap/tidb/pkg/testkit/testfailpoint"
 	"github.com/pingcap/tidb/tests/realtikvtest"
 	"github.com/stretchr/testify/require"
 	"github.com/tikv/client-go/v2/util"
@@ -73,17 +72,4 @@ func TestDDLTestEstimateTableRowSize(t *testing.T) {
 		size = ddl.EstimateTableRowSizeForTest(ctx, store, exec, partition)
 		require.Equal(t, 19, size)
 	}
-}
-
-func TestAddIndexValidateRangesFailed(t *testing.T) {
-	store := realtikvtest.CreateMockStoreAndSetup(t)
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("create table t (a int primary key, b int);")
-	tk.MustExec("insert into t values (1, 1);")
-
-	testfailpoint.Enable(t, "github.com/pingcap/tidb/pkg/ddl/loadTableRangesNoRetry", "return")
-	testfailpoint.Enable(t, "github.com/pingcap/tidb/pkg/ddl/validateAndFillRangesErr", "2*return")
-	tk.MustExec("alter table t add index idx(b);")
-	tk.MustExec("admin check table t;")
 }
